SHARMAN-3207: Telcovoice_manager is not running Intermittently
Reason for change:
In 64 bit mode , the dml function assings 8 byte value to a
4 byte integer which causes the memory corruption. The function
prototype must be changed to use proper data type to avoid this.
Test Procedure:
Telcovoice manager should not be crashed
Risks: Low

(cherry picked from commit a5ecc5e84e645381b6bb8cdec9fd2893d3115539)

SKYH4-7378: Compilation failure in telcovoice manager

Reason for change:
Compilation failure in telcovoice manager

Test Procedure: Build must be succeeded

Priority:P1
Risks: Medium

(cherry picked from commit 1f69db4d20aaf5a489b7747f4135c5169c589b60)

+++ source/TR-181/include/telcovoicemgr_services_apis_v2.h
@@ -3208,7 +3208,7 @@ ANSC_STATUS TelcoVoiceMgrDmlSetLinkState(TELCOVOICEMGR_VOICE_IP_LINK_STATE linkS
 ANSC_STATUS TelcoVoiceMgrDmlSetDirectoryNumber(UINT uiService, UINT uiProfile, UINT uiLine, char* pDirName);
 ANSC_STATUS TelcoVoiceMgrDmlSetLineEnable(UINT uiService, UINT uiProfile, UINT uiLine, BOOL LineEnable);
 ANSC_STATUS TelcoVoiceMgrDmlGetLineStatus(UINT uiService, UINT uiProfile, UINT uiLine, UINT *pLineStatus);
-ANSC_STATUS TelcoVoiceMgrDmlGetLineCallState(UINT uiService, UINT uiProfile, UINT uiLine, ULONG* pCallState);
+ANSC_STATUS TelcoVoiceMgrDmlGetLineCallState(UINT uiService, UINT uiProfile, UINT uiLine, UINT *pCallState);
 
 ANSC_STATUS TelcoVoiceMgrDmlSetOutboundProxy(UINT uiService, UINT uiProfile, char* pProxyName);
 ANSC_STATUS TelcoVoiceMgrDmlSetProxyServer(UINT uiService, UINT uiProfile, char* pProxyServer);
